decoupling knappsack uploading logic from the grunt plugin
