+++ package.json
@@ -40,9 +40,10 @@
     "gruntplugin"
   ],
   "dependencies": {
+    "form-data": "^0.1.2",
+    "lodash": "^2.4.1",
     "q": "^1.0.1",
-    "url": "^0.7.9",
     "request": "^2.34.0",
-    "form-data": "^0.1.2"
+    "url": "^0.7.9"
   }
 }
